W7 Revamp

+++ C++/week7_activity4.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-  for (float i = 1; i <= 10; i++){
+  for (int i = 1; i <= 10; i++){
     cout << i << "\t" << (i * 2) << "\t" << (i * 3) << "\t" << (i * 4) << "\t" << (i * 4) << "\t" << (i * 5) << "\t" << (i * 6) << "\t" << (i * 7) << "\t" << (i * 8) << "\t" << (i * 9) << "\t" << (i * 10) << endl;
   }
